.github/workflows/ci.yml: refactor for Poetry
This commit overhauls the CI/CD workflow to support the change to Poetry. It includes extensive linting and testing in a parallel fashion.

In addition, heavy caching is used to reduce the CI/CD job times. Lastly, CodeCov is available, but if the token is missing, the workflow will not fail.